requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Django==5.0.1
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-grappelli==3.0.8
django-jazzmin==2.6.0
djangorestframework==3.14.0
executing==2.0.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flat==0.3.2
greenlet==3.0.3
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2021.11.10
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.3
terregex==0.0.6
text-unidecode==1.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
yalm==0.0.2a7
